Split FixtureTest into partial files
